Add character chats subnav item
